# ita_root/ita_api_organization/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: ITA Version2.0 API
  description: ITAv2.0用API
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/class/info/:
    get:
      tags:
      - Conductor Controll
      description: ConductorClassの基本情報取得
      operationId: get_conductor_class_info
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_13'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/class/{conductor_class_id}/:
    get:
      tags:
      - Conductor Controll
      description: ConductorClassの情報取得
      operationId: get_conductor_class_data
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: conductor_class_id
        in: path
        description: Conductor Class ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_6'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/class/maintenance/:
    post:
      tags:
      - Conductor Controll
      description: Conductorの新規登録
      operationId: post_conductor_data
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: 登録するConductorの構成情報
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_12'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/class/maintenance/{conductor_class_id}/:
    patch:
      tags:
      - Conductor Controll
      description: Conductorの更新
      operationId: patch_conductor_data
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: conductor_class_id
        in: path
        description: Conductor Class ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: 更新するConductorの構成情報
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_12'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/execute/info/:
    get:
      tags:
      - Conductor Controll
      description: "Conductor,Operationのメニューの基本情報および項目情報を取得する"
      operationId: get_conductor_execute_info
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_14'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/execute/filter/{target}/:
    post:
      tags:
      - Conductor Controll
      description: "Conductor,Operationを対象に、検索条件を指定し、レコードを取得する"
      operationId: post_execute_filter
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: target
        in: path
        description: conductor_list or operation_list
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/filter_target_body'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_5'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/execute/filter/{target}/search/candidates/{column}/:
    get:
      tags:
      - Conductor Controll
      description: 表示フィルタで利用するプルダウン検索の候補一覧を取得する
      operationId: get_execute_search_candidates
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: target
        in: path
        description: conductor_list or operation_list
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: column
        in: path
        description: REST用項目名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_3'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/execute/class/info/:
    get:
      tags:
      - Conductor Controll
      description: ConductorClassの基本情報取得
      operationId: get_conductor_execute_class_info
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_13'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/execute/class/{conductor_class_id}/:
    get:
      tags:
      - Conductor Controll
      description: ConductorClassの情報取得
      operationId: get_conductor_execute_class_data
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: conductor_class_id
        in: path
        description: Conductor Class ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_6'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/execute/:
    post:
      tags:
      - Conductor Controll
      description: Conductor作業実行
      operationId: post_conductor_excecute
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/conductor_execute_body'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_12'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
  /api/{organization_id}/workspaces/{workspace_id}/ita/menu/{menu}/conductor/info/{conductor_instance_id}/:
    get:
      tags:
      - Conductor Controll
      description: Conductorの基本情報取得
      operationId: get_conductor_info
      parameters:
      - name: organization_id
        in: path
        description: OrganizationID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: workspace_id
        in: path
        description: WorkspaceID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: menu
        in: path
        description: メニュー名
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: conductor_instance_id
        in: path
        description: Conductor Instance ID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_13'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_TEMPLATE'
      x-openapi-router-controller: controllers.conductor_controll_controller
...
components:
  schemas:
...
    ERROR_TEMPLATE:
      type: object
      properties:
        result:
          type: string
          example: 001-002
        message:
          type: string
...
    inline_response_200_5:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: array
          items:
            $ref: '#/components/schemas/FILTER_PARAMETERS'
        message:
          type: string
      example:
        result: 000-00000
        data:
        - file:
            rest_key_name: base64 encode file
          parameter:
            rest_key_name: rest_key_name
        - file:
            rest_key_name: base64 encode file
          parameter:
            rest_key_name: rest_key_name
        message: message
...
    inline_response_200_6:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: object
        message:
          type: string
      example:
        result: 000-00000
        data: {}
        message: message
...
    inline_response_200_12:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          type: boolean
        message:
          type: string
      example:
        result: 000-00000
        data: true
        message: message
    inline_response_200_13:
      type: object
      properties:
        result:
          type: string
          example: 000-00000
        data:
          $ref: '#/components/schemas/inline_response_200_13_data'
        message:
          type: string
      example:
        result: 000-00000
        data:
          dict:
            node_status:
              status_id1: status_name1
              status_id2: status_name2
            node_type:
              node_type_id1: node_type_name1
              node_type_id2: node_type_name2
            conductor_status:
              status_id1: status_name1
              status_id2: status_name2
            movement:
              movement_id1: movement_name1
              movement_id2: movement_name2
            operation:
              operation_id1: operation_name1
              operation_id2: operation_name2
            conductor:
              conductor_id1: conductor_name1
              conductor_id2: conductor_name2
          list:
            node_status:
            - id: string
              name: string
            node_type:
            - id: string
              name: string
            conductor_status:
            - id: string
              name: string
            movement:
            - id: string
              name: string
            operation:
            - id: string
              name: string
            conductors:
            - id: string
              name: string
        message: message
...
    filter_target_body:
      properties:
        rest_key_name:
          $ref: '#/components/schemas/FILTER_REQUEST_PARAMETERS'
      description: 検索条件を設定
...